// virtual-playground/shared/types/SimulationBundle.ts
export interface LogicRule {
  id: string;
  inputPinOrKey: string;
  operator: '<' | '>' | '==' | '!=' | '<=' | '>=';
  threshold: number | string;
  outputPinOrKey: string;
  actionValue: boolean | number | string;
}